Made Initial component for WelcomeEmail
